commit 9
Today, as requested, jQuery was deleted and the logic was changed with a different approach. And completed the file path switch, dividing it into front-end and API folders according to requirements. Facilitating subsequent submissions. In addition, the English translation of the annotations has not been completed yet as the final check has not been done.